requirements.txt
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
